yazlab2-main-1/makale_proje/backend/app/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Werkzeug==2.0.1
python-dotenv==0.19.0
PyJWT==2.1.0
python-magic==0.4.27
PyPDF2==3.0.1
python-dateutil==2.8.2
pdfminer.six==20221105
PyMuPDF==1.23.8
Pillow==10.2.0
spacy>=3.0.0
flask-cors==3.0.10
cryptography>=36.0.0
 